fixes to align data with new speeduino data formatting and name changes

+++ GPIO MINI/GPIO_MINI_V0_003/mini_gpiov0_003/utils.h
@@ -6,7 +6,7 @@ These are some utility functions and variables used through the main code
 
 #include <Arduino.h>
 
-int freeRam ();
+uint16_t freeRam ();
 void setPinMapping(byte boardID);
 
 //unsigned int PW();
